requirements.txt
streamlit>=1.42
pandas>=2.2
openpyxl>=3.1
rapidfuzz>=3.9
PyYAML>=6.0
requests>=2.32
